Preconditions to constructors added

+++ src/com/codelemma/finances/Utils.java
@@ -40,21 +40,19 @@ public class Utils {
 	}
 	
 	static public boolean alertIfNotInBounds(Context context, String fieldData, int min, int max, String fieldName) {
-	    if (Double.parseDouble(fieldData) < min) {
+	    if (Integer.parseInt(fieldData) < min) {
 	    	new AlertDialog.Builder(context).setTitle("\""+ fieldName +"\" too low")
 	        	                             .setMessage("Please, fill in the field with value higher than " +min+".")
 	        	                             .setNeutralButton("Close", null)
 	        	                             .show();
 	    	return true;
-	    } else if(Double.parseDouble(fieldData) > max) {
+	    } else if(Integer.parseInt(fieldData) > max) {
 	    	new AlertDialog.Builder(context).setTitle("\""+ fieldName +"\" too high")
             .setMessage("Please, fill in the field with value lower than " +max+".")
             .setNeutralButton("Close", null)
             .show();
-return true;
-	    	
+            return true;
 	    }
 	    return false;
 	}
-
 }
